+++ services/ui-src/src/types/users.ts
@@ -25,9 +25,7 @@ export interface MFPUser {
 }
 
 export interface UserContextShape {
-  user: MFPUser | null;
   logout: () => Promise<void>;
-  showLocalLogins: boolean | undefined;
   loginWithIDM: () => void;
   updateTimeout: () => void;
   getExpiration: Function;
