+++ services/ui-src/src/utils/auth/UserProvider.tsx
@@ -1,39 +1,62 @@
-import { createContext, ReactNode, useCallback, useEffect } from "react";
+import {
+  createContext,
+  ReactNode,
+  useCallback,
+  useEffect,
+  useMemo,
+} from "react";
 import { useLocation } from "react-router-dom";
 import { Auth } from "aws-amplify";
 import config from "config";
-import { initAuthManager, useStore } from "utils";
+import { initAuthManager, updateTimeout, getExpiration, useStore } from "utils";
 import { PRODUCTION_HOST_DOMAIN } from "../../constants";
 
 import { MFPUser, UserContextShape, UserRoles } from "types/users";
 
 export const UserContext = createContext<UserContextShape>({
-  user: null,
-  showLocalLogins: undefined,
   logout: async () => {},
   loginWithIDM: () => {},
   updateTimeout: () => {},
   getExpiration: () => {},
 });
 
+const authenticateWithIDM = async () => {
+  const authConfig = Auth.configure();
+  if (authConfig?.oauth) {
+    const oAuthOpts = authConfig.oauth;
+    const domain = oAuthOpts.domain;
+    const responseType = oAuthOpts.responseType;
+    const redirectSignIn = (oAuthOpts as any).redirectSignIn;
+    const clientId = authConfig.userPoolWebClientId;
+    const url = `https://${domain}/oauth2/authorize?identity_provider=Okta&redirect_uri=${redirectSignIn}&response_type=${responseType}&client_id=${clientId}`;
+    window.location.assign(url);
+  }
+  const cognitoHostedUrl = new URL(
+    `https://${config.cognito.APP_CLIENT_DOMAIN}/oauth2/authorize?identity_provider=${config.cognito.COGNITO_IDP_NAME}&redirect_uri=${config.APPLICATION_ENDPOINT}&response_type=CODE&client_id=${config.cognito.APP_CLIENT_ID}&scope=email openid profile`
+  );
+  window.location.replace(cognitoHostedUrl);
+};
+
 export const UserProvider = ({ children }: Props) => {
   const location = useLocation();
   const isProduction = window.location.origin.includes(PRODUCTION_HOST_DOMAIN);
 
   // state management
-  const {
-    user,
-    showLocalLogins,
-    setUser,
-    setShowLocalLogins,
-    logout,
-    loginWithIDM,
-    updateTimeout,
-    getExpiration,
-  } = useStore();
+  const { user, showLocalLogins, setUser, setShowLocalLogins } = useStore();
 
   // initialize the authentication manager that oversees timeouts
   initAuthManager();
+
+  const logout = async () => {
+    try {
+      setUser(null);
+      await Auth.signOut();
+      localStorage.clear();
+    } catch (error) {
+      console.log(error); // eslint-disable-line no-console
+    }
+    window.location.href = config.POST_SIGNOUT_REDIRECT;
+  };
 
   const checkAuthState = useCallback(async () => {
     // Allow Post Logout flow alongside user login flow
@@ -70,7 +93,7 @@ export const UserProvider = ({ children }: Props) => {
       setUser(currentUser);
     } catch (error) {
       if (isProduction) {
-        loginWithIDM();
+        authenticateWithIDM();
       } else {
         setShowLocalLogins(true);
       }
@@ -100,14 +123,17 @@ export const UserProvider = ({ children }: Props) => {
     checkAuthState();
   }, [location, checkAuthState]);
 
-  const values: UserContextShape = {
-    user,
-    logout,
-    showLocalLogins,
-    loginWithIDM,
-    updateTimeout,
-    getExpiration,
-  };
+  const values: UserContextShape = useMemo(
+    () => ({
+      user,
+      logout,
+      showLocalLogins,
+      loginWithIDM: authenticateWithIDM,
+      updateTimeout,
+      getExpiration,
+    }),
+    [user, logout, showLocalLogins]
+  );
 
   return <UserContext.Provider value={values}>{children}</UserContext.Provider>;
 };
