+++ services/ui-src/src/utils/state/useStore.ts
@@ -1,7 +1,5 @@
 import { create } from "zustand";
 import { devtools, persist } from "zustand/middleware";
-import { Auth } from "aws-amplify";
-import config from "config";
 // types
 import {
   MfpUserState,
@@ -12,7 +10,6 @@ import {
   MfpReportState,
   ReportMetadataShape,
 } from "types";
-import { getExpiration, updateTimeout } from "utils/auth/authLifecycle";
 
 // USER STORE
 const userStore = (set: Function) => ({
@@ -26,35 +23,6 @@ const userStore = (set: Function) => ({
   // toggle show local logins (dev only)
   setShowLocalLogins: () =>
     set(() => ({ showLocalLogins: true }), false, { type: "showLocalLogins" }),
-  // context actions
-  logout: async () => {
-    try {
-      useStore().setUser(null);
-      await Auth.signOut();
-      localStorage.clear();
-    } catch (error) {
-      console.log(error); // eslint-disable-line no-console
-    }
-    window.location.href = config.POST_SIGNOUT_REDIRECT;
-  },
-  loginWithIDM: async () => {
-    const authConfig = Auth.configure();
-    if (authConfig?.oauth) {
-      const oAuthOpts = authConfig.oauth;
-      const domain = oAuthOpts.domain;
-      const responseType = oAuthOpts.responseType;
-      const redirectSignIn = (oAuthOpts as any).redirectSignIn;
-      const clientId = authConfig.userPoolWebClientId;
-      const url = `https://${domain}/oauth2/authorize?identity_provider=Okta&redirect_uri=${redirectSignIn}&response_type=${responseType}&client_id=${clientId}`;
-      window.location.assign(url);
-    }
-    const cognitoHostedUrl = new URL(
-      `https://${config.cognito.APP_CLIENT_DOMAIN}/oauth2/authorize?identity_provider=${config.cognito.COGNITO_IDP_NAME}&redirect_uri=${config.APPLICATION_ENDPOINT}&response_type=CODE&client_id=${config.cognito.APP_CLIENT_ID}&scope=email openid profile`
-    );
-    window.location.replace(cognitoHostedUrl);
-  },
-  updateTimeout: updateTimeout,
-  getExpiration: getExpiration,
 });
 
 // BANNER STORE
